Update README.md
The files implement the methods developed in the paper “Robust Data Driven Iterative Learning Control for Linear-Time-Invariant and Hammerstein-Wiener Systems”, authored by Jianfei Dong, which has been accepted for publication in IEEE TCYB.

The main file of the demo program is “demo_rddilc_hamwien_siso_MC.m.m”, which calls the other (precomplied for fast running) functions in the same folder. Please note that these are Monte Carlo simulations for a few hundred times. Running it will take some time. The simulations are based on randomly generated stochastic disturbances. Therefore, the results may not be exactly the same with those written in the paper. 

All the codes are developed in MATLAB R2014b, and shall be put into the same folder.

For suggestions or requests, please contact with Jianfei Dong: jfeidong@hotmail.com 

Copyright: Jianfei Dong (C), on Apr. 26, 2021.